Implement listening comprehension mode
Added sound files for pronunciation of foreign words in Klingon.

Added helper function and implement feature so that the sound is played when the player clicks on a cell in the grid.

Sound does NOT play when an action button is selected or when clicking on a cell to deselect it.